Fix type error through type coercion.
This is necessary because the Futhark type checker does not understand
arithmetic. Sizes are *uninterpreted* terms, and are considered purely
symbolically/syntactically. (The reasons for this are somewhat subtle,
but it's not _just_ the case that we haven't gotten around to it -
there are good reasons not to just do it in all cases.)

Also, there is no way to explicitly apply type parameters, but you can
always use a type ascription (not a coercion) to make them what you
want. This is very verbose, but perhaps OK for code generation.