Rewriting beginning of Section 5.
This is where my explanation of reverse AD was wrong.
I have rewritten it, and I have added a picture that hopefully
should clarify the process.

Also, I have starts `log.md` that contains big changes that we
did to the paper.  Hopefully, this will make writing the final
response easier.